testQues/users/serializers.py
from rest_framework import serializers

from notification_service.models import Client


class ClientSerializer(serializers.ModelSerializer):
    class Meta:
        model = Client
        fields = "__all__"
