testQues/requirements.txt
Faker~=18.9.0
Django~=4.2.1
djangorestframework~=3.14.0
apscheduler
drf_yasg
django_q